refactored test files as well to new packaging
