+++ applications/track-manager/src/test/scala/com/hashmapinc/tempus/edge/track/manager/DriverTest.scala
@@ -1,4 +1,4 @@
-package com.hashmapinc.tempus.edge.track_manager
+package com.hashmapinc.tempus.edge.track.manager
 
 import org.scalatest.FlatSpec
